requirements.txt
transformers
gradio
python-docx
pdfplumber
markdown